Finished DB creation for initial feature
REMOVED: fx_initCL.lua -- JINK
ADDED: Dump function for table debugging --JINK
ADDED: DrawTxt function to draw the LSC text on screen -- JINK
ADDED: SetVehicleOutsideGarage / SetVehicleInGarage functions. They will set the player in the LSC locations if they are in a vehicle and within a distance and press the HOME key. Once they trigger the close menu button they will be set back outside. The menu currently gives and error when closing due to the way the close() of ft_menuBuilder not receiving the correct data but still closes the window. -- JINK
ADDED: General DB additions to prepare for the vehicle mods. Hopefully this will be all the columns we need for vehicle mods.-- JINK
UPDATE: Updated Supertables to latest version --JINK