Database Commit and Repair Commit
ADDED:  Server event to add the users car in the DB if they own it. Only when they enter the customs the shop for the first time with that vehicle. -- JINK
ADDED: Server event for confirming the repair of a vehicle and it will debit the users account for the price of the repair. -- JINK
CLEANUP: Cleaned up whitespace and general other commits. -- JINK