Pushing towards alpha complete
ADDED: Windowtints are able to be applied via the menu now -- JINK
ADDED: Neons can be added in the following configurations (front only, left and right, back only, All sides) with a variety of colors. --JINK
ADDED: Cars can be painted now via a primary color, secondary color, pearlescent color, and rims can be painted the same way. --JINK
UPDATE: Config.lua is updated to almost a complete state. There are still a few mods that need to be added (Spoilers and such) but those will be added once the current mods are completed and able to be installed. --JINK
UPDATE: Moved motype,mod,wheeltype etc from server.lua to the client.lua via data on the menu to clean up the code and be more effient. ... .. YAY EFFICIENCY! --JINK