Cleaning up Bugs/working towards Alpha
ADDED: Menu.lua Function to add tables of buttons to a menu. This works better for what my idea for FX_Customs will need. --JINK
ADDED: Client.lua Vehicle generation thread to make testing easier as my trainer isn't working for vehicle spawning. This will be removed after completed. --JINK
ADDED: Client.lua A check for if the player is on a bike AKA motorcycle and this will change what option are presented in the FX customs shops. --JINK
UPDATE: Client.lua Moved Setting mod checks around. Nested Neons and Smoke to make it so that either sets smoke RGB or Neon RGB instead of both to prevent a bug that was identified during testing. --JINK
UPDATE: Client.lua Wheel check as I eventually will break wheel type out to a separate menu and then rims into their own category. --JINK
UPDATE: Client.lua Turbo is now a toggle like it needs to be as it wasn't actually doing anything but debiting the money from the user. --JINK
FIX: Fixed Issue #1 identified by JINK. The issue was with if painCar ~= nil. We had to add colrIndex ~= nil because when setting vehicle mods it was triggering car painting. AKA resetting all of the settings to -1. --JINK