Vehicle Extras Auto Added / Client comments and cleanup
ADDED: Vehicle extras are now auto added to the vehicle shop menu based on the amount of extras a vehicle can have. They can be enabled or disabled depening on what the user wants. This menu is dynamic and checks up to 100 values, which is excessive and should probably be 20. --JINK
ADDED: Comments for general code cleanup and to prevent console spam cleaned out a few Trace debugs. --JINK
ADDED: GetVehicleData() function. This is huge and will prevent a lot of repeat work for me in the future for getting vehicle data. I'm an idiot but i'm learning. --JINK
UPDATE: Server.lua is being updated to hopefully limit the ability to double purchase mods. As setting wheeltypes twice will crash the client if its the same wheel. AddVehicle is being updated to add vehicle data to the DB just needs to be more current. --JINK
UPDATE: Server.lua Notify is being updated also to make it more appealing. --JINK
UPDATE: Config.lua to reflect the new extras menus. 
UPDATE: Config.lua found missing commas that didn't affect the script but for my muscle memory i need to correct those things. --JINK