DB is being worked on.
ADDED: Server.lua Almost if not all of the mods should now be correctly added to the DB columns. The car is added to the DB if it's not stolen. All the mods in the DB should be updated when you send the query. --JINK
ADDED: Server.lua split AddVehicle() into two parts from the setting vehicle data to try to fix the bug with updating vehicle data. Currently it did not fix it and I'm not really in the mood to look further tonight. --JINK
ADDED: Server.lua repair should now correctly notify if you if you did not have enough cash for the repair.
ADDED: Client.lua added more information to the vehicle data function. This is currently storing a lot of vehicle information in the table. Should be pretty useful for future updates to the script. --JINK
BUG: Currently the update query takes two submits to update the data for the button press. IE updating the transmission to 3 takes two presses for it to show in the DB. It's one press behind basically. Need to find the source of this. --JINK