Final Commit Priot to 1.0 Beta Release
This wont be as detailed of a commit but basically this will most likely be the last commit prior to beta release of version 1.0. There are a few more bugs that need to be worked out before I'm comfortable releasing this as a beta product since this will be the first public work I release and my first thing i've programmed.

ADDED: Server.lua will have the bulk of this commit. It should allow for the buying of all mods and disable you from double buying most of them. There most likely will need to be a delay or wait in the query to prevent spamming.  --JINK
BUG: Turbo is either not working or not notifying you of being bought. --JINK
BUG: Neons once bought the first time have some sort of delay or lag and weirdness where they feel unresponsive for buying the rest.--JINK
BUG: Headlight currently show no notification once bought but still will set. So i'm unsure if they are preventing double buying or not. --JINK
BUG: Neon/TireSmoke are linked somehow and once you buy a neon color the tire smoke color is linked to that RGB so even though you select red smoke you just wont be able to buy the color your neons are. --JINK
BUG: Horns work but the truck horn is able to bought unlimited amount of times. I'm assuming to the value being false. --JINK