Essential Mode Version Commit
This brought about a bunch of changes in this commit. Essential mode released a plugin for sql support. I've done a bit work to make this conversion work but hopefully it's on the same level as the supertables version. I think unless supertables gets some more updates this will be the main branch with supertables being worked on when i get time. --JINK