Preview Mod Feature Started
Removed: Ft_menuBuilder from resource. Now users will get from the creators resource as edits are included in the creators resource. --JINK
ADDED: Config.lua hoverEvent was added to some buttons that triggers setting the mods. This needs to be tweaked around because once the mod is set the user can just leave. Maybe make the user save the cost in the table to pay for the price or set the user back to the state they were in when they came. However once they buy a mod it should stay. I need to figure out the logistics of those. Once that's figured out this should be considered feature complete and pushed to public. As doing it for the other mods or future mods should be simple. --JINK