Building on preview functions
MODIFIED: client.lua A lot of changes have happened. I've moved a lot of the guts of the events to actual functions now. This should make future improvements easier for me. This was a lot of work to make sure everything worked. Also there may be a lot of redundancy going on here so once preview mode is done. I'll look at that and pick out things that don't need to be done or queries that should be done. --JINK

ADDED: Client.lua i've started adding in the resetting of previews to original state in the db if it doesn't match. This just needs to be done for all the preview mods in a good way. --JINK

ADDED: Config.lua added Hover events to config that will trigger the resetting of mods. This should fix the problem that stumped me for a few days. --JINK

MODIFIED: Server.lua to only add the car on first entrance to the garage. Then each mod should individually update the db once it's paid for and changed. Also a big change for me from the previous way because this prevents the preview mods from being submitted to the DB. --JINK