Completion of basic preview feature.
Feature for preview mode is basically complete. Hopefully this works as expected as I was not able to test every case and rewrote a lot to make this work. --JINK

BUGFIX: Setting paint for all vehicles should be fixed as now we check and compare the plates in the database before setting the paint. --JINK
BUGFIX: Truck horn and other "False" mods should not be able to be bought infintely now. --JINK